refactor: :recycle: move BuildStyle to internals
