+++ src/seedcase_flower/__init__.py
@@ -1,6 +1,7 @@
 """Module containing all source code."""
 
 from .cli import build, view
-from .config import BuildStyle, Config
+from .config import Config
+from .internals import BuildStyle
 
 __all__ = ["build", "view", "Config", "BuildStyle"]
