+++ src/seedcase_flower/cli.py
@@ -1,18 +1,9 @@
 """Functions for the exposed CLI."""
 
-from enum import Enum
 from pathlib import Path
 from typing import Any, Optional
 
-from seedcase_flower.internals import _read_properties, _resolve_uri
-
-
-class BuildStyle(Enum):
-    """Built-in styles for outputting to file."""
-
-    quarto_one_page = "quarto_one_page"
-    quarto_resource_listing = "quarto_resource_listing"
-    quarto_resource_tables = "quarto_resource_tables"
+from seedcase_flower.internals import BuildStyle, _read_properties, _resolve_uri
 
 
 def build(
