+++ src/seedcase_flower/internals.py
@@ -1,8 +1,17 @@
 """Helper functions for private use."""
 
 import json
+from enum import Enum
 from pathlib import Path
 from typing import Any
+
+
+class BuildStyle(Enum):
+    """Built-in styles for outputting to file."""
+
+    quarto_one_page = "quarto_one_page"
+    quarto_resource_listing = "quarto_resource_listing"
+    quarto_resource_tables = "quarto_resource_tables"
 
 
 # TODO Extend to parse strings and return either URL or Path
