+++ tools/vulture-allowlist.py
@@ -1,8 +1,5 @@
 # ruff: noqa
 # mypy: ignore-errors
-quarto_one_page  # unused variable (src/seedcase_flower/cli.py:13)
-quarto_resource_listing  # unused variable (src/seedcase_flower/cli.py:14)
-quarto_resource_tables  # unused variable (src/seedcase_flower/cli.py:15)
-quarto_resource_listing  # unused variable (src/seedcase_flower/config.py:13)
-quarto_resource_tables  # unused variable (src/seedcase_flower/config.py:14)
-_style_with_template_dir  # unused method (src/seedcase_flower/config.py:65)
+quarto_resource_listing  # unused variable (src/seedcase_flower/internals.py:13)
+quarto_resource_tables  # unused variable (src/seedcase_flower/internals.py:14)
+_style_with_template_dir  # unused method (src/seedcase_flower/config.py:56)
